requirements.txt
streamlit
pandas
supabase
google-generativeai
plotly
fpdf